refactor: Modified, removed or commented out code from old schema and auth
