+++ src/lib/auth.ts
@@ -1,25 +1,25 @@
-import { getSession } from "@auth0/nextjs-auth0";
-import { Role } from "@prisma/client";
-import { redirect } from "next/navigation";
+// import { getSession } from "@auth0/nextjs-auth0";
+// import { Role } from "@prisma/client";
+// import { redirect } from "next/navigation";
 
-import { db } from "~/server/db";
+// import { db } from "~/server/db";
 
-export async function checkAdmin() {
-  const session = await getSession();
+// export async function checkAdmin() {
+//   const session = await getSession();
 
-  if (!session?.user) {
-    redirect("/api/auth/login");
-  }
+//   if (!session?.user) {
+//     redirect("/api/auth/login");
+//   }
 
-  const user = await db.user.findUnique({
-    where: {
-      email: session.user.email as string,
-    },
-  });
+//   const user = await db.user.findUnique({
+//     where: {
+//       email: session.user.email as string,
+//     },
+//   });
 
-  if (user?.role !== Role.ADMIN) {
-    redirect("/unauthorized");
-  }
+//   if (user?.role !== Role.ADMIN) {
+//     redirect("/unauthorized");
+//   }
 
-  return;
-}
+//   return;
+// }
